+++ lib/pocket/version.rb
@@ -1,3 +1,3 @@
 module Pocket
-  VERSION = "0.6.2"
+  VERSION = "0.6.3"
 end
